Making changes to prevent circular import
