+++ climakitae/explore/warming.py
@@ -17,7 +17,6 @@ from climakitae.util.utils import (
     scenario_to_experiment_id,
     timescale_to_table_id,
     resolution_to_gridlabel,
-    drop_invalid_wl_sims,
 )
 
 from tqdm.auto import tqdm
@@ -89,8 +88,8 @@ class WarmingLevels:
         self.catalog_data = self.catalog_data.stack(all_sims=["simulation", "scenario"])
 
         # Dropping invalid simulations that come up from stacking scenarios and simulations together
-        self.catalog_data = drop_invalid_wl_sims(
-            self.catalog_data, self.wl_params.downscaling_method
+        self.catalog_data = _drop_invalid_wl_sims(
+            self.catalog_data, self.wl_params
         )
 
         if self.wl_params.anom == "Yes":
@@ -338,57 +337,36 @@ class WarmingLevelChoose(DataParameters):
             self.anom = "Yes"
 
 
-def drop_invalid_wrf_sims(ds):
+def _drop_invalid_sims(ds, selections):
     """
-    Drop invalid WRF simulations from the given dataset since there is an unequal number of simulations per SSP.
+    As part of the warming levels calculation, the data is stacked by simulation and scenario, creating some empty values for that coordinate.
+    Here, we remove those empty coordinate values.
 
     Parameters
     ----------
     ds : xr.Dataset
-        The dataset containing WRF simulations. The dataset must have a
+        The dataset must have a
         dimension `all_sims` that results from stacking `simulation` and
         `scenario`.
+    data_catalog: pd.DataFrame
+        intake catalog, loaded as a pandas dataframe
 
     Returns
     -------
     xr.Dataset
-        The dataset with only valid WRF simulations retained.
+        The dataset with only valid simulations retained.
 
     Raises
     ------
     AttributeError
         If the dataset does not have an `all_sims` dimension.
-
-    Notes
-    -----
-    - For datasets with a resolution of '3 km', no simulations are dropped, and the original dataset is returned.
-    - For datasets with a resolution of '9 km' at hourly timescale, only 10 simulations are returned.
-    - For datasets with a resolution of '9 km' at daily/monthly timescale, only 6 simulations are returned.
-    - For datasets with a resolution of '45 km' at hourly timescale, only 7 simulations are returned.
-    - For datasets with a resolution of '45 km' at daily/monthly timescale, only 6 simulations are returned.
     """
-    if "all_sims" not in ds.dims:
-        raise AttributeError(
-            "Missing an `all_sims` dimension on the dataset. Create `all_sims` with .stack on `simulation` and `scenario`."
-        )
-
-    # Checking for derived variables separately since we don't store their IDs in the catalog
-    # Future derived variables that don't use `t2` will be broken because of this function.
-    variable = ds.variable_id
-    if "derived" in variable:
-        variable = "t2"
-
-    # Find valid simulation from catalog
-    df = intake.open_esm_datastore(data_catalog_url).df
-    filter_df = df[
-        (df["activity_id"] == "WRF")
-        & (df["table_id"] == timescale_to_table_id(ds.frequency))
-        & (df["grid_label"] == resolution_to_gridlabel(ds.resolution))
-        & (df["variable_id"] == variable)
-        & (df["experiment_id"] != "historical")
-        & (df["experiment_id"] != "reanalysis")
-        & (df["source_id"] != "ensmean")
-    ]
+    df = get_cat_subset(selections).df
+    
+    # Just trying to see simulations across SSPs, not including historical period
+    filter_df = df[df['experiment_id'] != 'historical']
+    
+    # Creating a valid simulation list to filter the original dataset from
     valid_sim_list = list(
         zip(
             filter_df["activity_id"]
